requirements.txt
pandas
python-dotenv
pinecone-client
streamlit
langchain
langchain-community
langchain-openai
langchain-core